// src/types/api-requests.ts
interface ResponseBody {
  success: boolean;
  message?: string;
  token?: string;
  role?: string;
  [key: string]: unknown;
}
// ...
interface CategoriesRequestBody {
  page?: number;
  resultsOnPage?: number;
  name_uk?: string;
  name_en?: string;
  target?: "" | "men" | "women" | "kids";
  sortField?: string;
  sortDesc?: boolean;
}
interface CategoriesResponseBody extends ResponseBody {
  categories?: Category[];
  count?: number;
}
interface AddCategoryRequestBody {
  category: Category;
}
interface UpdateCategoryRequestBody {
  id: string;
  newName_uk: string;
  newName_en: string;
  newTarget: "" | "men" | "women" | "kids";
}
interface RemoveCategoryRequestBody {
  id: string;
}